Move `Kernel::adoptSync()` and `executeSync()` into separate `KernelSync` interface.
